caldown/backend/requirements.txt
flask
mysql-connector-python
PyJWT
requests
uwsgi